se implemento correctamente los cambios

+++ StoreCollection/src/main/java/com/proyecto/StoreCollection/dto/request/AtributoRequest.java
@@ -7,5 +7,5 @@ import lombok.Data;
 public class AtributoRequest {
     @NotBlank
     private String nombre;
-    private Long tiendaId;
+    private Integer tiendaId;
 }
